+++ src/jdom/jacdac-jdom.ts
@@ -40,3 +40,7 @@ export * from "./flags"
 export * from "./settingsclient"
 export * from "./pack"
 export * from "./light"
+export * from "./registerhost"
+export * from "./servicehost"
+export * from "./devicehost"
+export * from "./controlservicehost"
